Some changes in read me, and add comments
