+++ eventhubs-streamanalytics-eventhubs/tools/README.md
@@ -9,10 +9,10 @@ extensions:
 - platforms: dotnet
 ---
 
-# Getting started on processing received message from eventhub #
+# Getting started on processing received message from Event Hubs #
 
-Azure eventhubs-streamanalytics-eventhubs sample for processing received message from eventhub -
- - Create an EventHubConsumerClient based on connection string for the `Listen` or `Manage` Shared Access Policy for the eventhub.
+Azure eventhubs-streamanalytics-eventhubs sample for processing received message from Event Hubs -
+ - Create an EventHubConsumerClient based on Event Hub-level connection string.
  - Read events from EventHubConsumerClient.
  - Processing received message and write to a .csv file.
 
@@ -21,10 +21,15 @@ To run this sample:
 
 If you don't have an Azure subscription, create a [free account] before you begin.
 	
-Go to [Azure portal], sign in with your account. You can find how to [create event hub service] from here. 
+Go to [Azure portal], sign in with your account. You can find how to [create event hub] from here. 
 
-Get the EventHubConnectionString for the `Listen` or `Manage` Shared Access Policy for the eventhub.
+After creating an event hub within the Event Hubs namespace, you can get the Event Hub-level EventHubConnectionString with following steps.
+	1. In the list of event hubs, select your event hub.
+	2. On the Event Hubs Instance page, select Shared Access Policies on the left menu.
+	3. Add a policy with appropriate permissions(Listen or Manage).
+	4. Click on added policy.
 
+Run the following command.
 ```bash	
 git clone https://github.com/Azure-Samples/streaming-at-scale.git
 ```
@@ -33,9 +38,9 @@ From within the `eh-asa-perform` folder run:
 ```bash
 dotnet run -- -c "<EventHubConnectionString>"
 ```
-The application will read from the eventhub and will measure latency (difference from Enqueued and Processed time) for each batch of recevied messages.
+The application will read from the event hub and will measure latency (difference from Enqueued and Processed time) for each batch of recevied messages.
 
 <!-- LINKS -->
 [free account]: https://azure.microsoft.com/free/?WT.mc_id=A261C142F
 [Azure portal]: https://ms.portal.azure.com/#home
-[create event hub service]: https://docs.microsoft.com/en-us/azure/event-hubs/event-hubs-create
+[create event hub]: https://docs.microsoft.com/en-us/azure/event-hubs/event-hubs-create
